refactor(account-factory): simplify state and introduce master account (#1612)
* simplify account-factory state

* do not introduce blank lines (keeping diffs small)

* fix(account-factory): separate key existence check from last-key guard

Split the combined condition into two distinct checks: first verify the
key exists, then ensure it isn't the user's last key. This gives callers
a clear error message instead of silently succeeding when deleting a
non-existent key.


* empty commit to trigger CI

* refactor(account-factory): convert USERS to IndexedMap

Replace manual secondary indexes (USERS_BY_KEY Set, USER_INDEXES_BY_NAME
Map) with a single IndexedMap that automatically maintains both indexes
via multi-value indexers (MultiIndex::new2, UniqueIndex::new2).


* feat(upgrade): add user state migration to consolidated User struct

Migrate legacy per-user storage maps (KEYS, USER_NAMES_BY_INDEX,
ACCOUNTS_BY_USER, ACCOUNT_COUNT_BY_USER, USER_INDEXES_BY_NAME) into the
new USERS IndexedMap<UserIndex, User>. Runs after delete_multisig so it
can read the already-migrated ACCOUNTS format.


---------

+++ dango/account-factory/src/lib.rs
@@ -6,4 +6,4 @@ mod state;
 pub use {account_querier::*, execute::*, query::*, state::*};
 
 /// The maximum number of accounts a user can be associated with.
-pub const MAX_ACCOUNTS_PER_USER: u8 = 5;
+pub const MAX_ACCOUNTS_PER_USER: usize = 5;
